requirements.txt
python-dotenv #used to call environment variables
jupyter 
pandas 
numpy
scikit-learn
pymongo
certifi
scipy
dill
pyaml
#-e . 